organize Foundry setup into src and tests
