+++ scripts/setup-foundry/src/commands/__init__.py
