+++ scripts/setup-foundry/src/config.py
