+++ scripts/setup-foundry/src/installation.py
