+++ scripts/setup-foundry/src/releases.py
