+++ scripts/setup-foundry/src/toolchain.py
