+++ scripts/setup-foundry/tests/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Foundry setup command-handler tests."""
